# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: 3rd Library API
  version: 1.0.0
  description: HTTP API for the backend.
servers:
  - url: /
...
paths:
...
  /api/articles:
    get:
      tags: [Articles]
      summary: List articles
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
        - name: favorited
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticlesResponse"
    post:
      tags: [Articles]
      summary: Create article
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "200":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/articles/feed:
    get:
      tags: [Articles]
      summary: List articles from followed users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            format: int64
            default: 0
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticlesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/articles/{slug}:
    get:
      tags: [Articles]
      summary: Get article
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Articles]
      summary: Update article
      parameters:
        - $ref: "#/components/parameters/Slug"
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleRequest"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      tags: [Articles]
      summary: Delete article
      parameters:
        - $ref: "#/components/parameters/Slug"
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Article deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/articles/{slug}/favorite:
    post:
      tags: [Articles]
      summary: Favorite article
      parameters:
        - $ref: "#/components/parameters/Slug"
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Favorited article
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Articles]
      summary: Unfavorite article
      parameters:
        - $ref: "#/components/parameters/Slug"
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Unfavorited article
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Use format: Token <jwt>"
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Token
    Forbidden:
      description: Forbidden
    NotFound:
      description: Not found
    UnprocessableEntity:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsResponse"
  schemas:
    ErrorsResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Profile:
      type: object
      required: [username, bio, image, following]
      properties:
        username:
          type: string
        bio:
          type: string
        image:
          type: string
          nullable: true
          description: profile picture id
          format: uuid
        following:
          type: boolean
...
    ArticleResponse:
      type: object
      required: [article]
      properties:
        article:
          $ref: "#/components/schemas/Article"
    ArticlesResponse:
      type: object
      required: [articles, articlesCount]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        articlesCount:
          type: integer
          format: int64
    Article:
      type: object
      required:
        [
          slug,
          title,
          description,
          body,
          tagList,
          createdAt,
          updatedAt,
          favorited,
          favoritesCount,
          author,
        ]
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        body:
          type: string
        tagList:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        favorited:
          type: boolean
        favoritesCount:
          type: integer
          format: int64
        author:
          $ref: "#/components/schemas/Profile"
    CreateArticleRequest:
      type: object
      required: [article]
      properties:
        article:
          type: object
          required: [title, description, body, tagList]
          properties:
            title:
              type: string
            description:
              type: string
            body:
              type: string
            tagList:
              type: array
              items:
                type: string
    UpdateArticleRequest:
      type: object
      required: [article]
      properties:
        article:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            body:
              type: string
...